Refactored the server to use more OO services, integrated plugin service
