--- Intersect.Server/Core/ServerLogic.cs
@@ -1,20 +1,0 @@
-﻿using Intersect.Threading;
-
-namespace Intersect.Server.Core
-{
-
-    internal sealed class ServerLogic : Threaded
-    {
-
-        public ServerLogic() : base("ServerLogic")
-        {
-        }
-
-        protected override void ThreadStart()
-        {
-            ServerLoop.RunServerLoop();
-        }
-
-    }
-
-}
