+++ Intersect.Server/Database/DbInterface.cs
@@ -1,4 +1,4 @@
-using System;
+﻿using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Collections.Immutable;
@@ -168,7 +168,7 @@ namespace Intersect.Server.Database
         }
 
         // Database setup, version checking
-        internal static bool InitDatabase([NotNull] ServerContext serverContext)
+        internal static bool InitDatabase([NotNull] IServerContext serverContext)
         {
             sGameDb = new GameContext(
                 CreateConnectionStringBuilder(Options.GameDb ?? throw new InvalidOperationException(), GameDbFilename),
@@ -1425,9 +1425,9 @@ namespace Intersect.Server.Database
                         gameDbLogger?.Debug($"Save took {elapsedMs}ms, {--gameSavesWaiting} saves queued.");
                         gameDbTraces.TryDequeue(out _);
                     }
-                    catch (Exception ex)
+                    catch (Exception exception)
                     {
-                        if (ex is InvalidOperationException)
+                        if (exception is InvalidOperationException)
                         {
                             //Collection was modified?
                             //Loop and try to save again!
@@ -1438,14 +1438,10 @@ namespace Intersect.Server.Database
                                     "Error Saving Game Database! Server will shutdown in order to prevent potential rollback scenarios!"
                                 );
 
-                                Task.Factory.StartNew(
-                                    () => Bootstrapper.OnUnhandledException(
-                                        Thread.CurrentThread.Name, new UnhandledExceptionEventArgs(ex, true)
-                                    )
-                                );
+                                ServerContext.DispatchUnhandledException(exception);
 
                                 gameDbLogger?.Error(
-                                    ex,
+                                    exception,
                                     "Error Saving Game Database! Server will shutdown in order to prevent potential rollback scenarios! [Failures: " +
                                     failures +
                                     "]"
@@ -1457,7 +1453,7 @@ namespace Intersect.Server.Database
                             //This should be a warning but I want to actually see it working in a real environment without people needing to change their configs for a few builds
                             //TODO change to .Warn
                             gameDbLogger?.Error(
-                                ex,
+                                exception,
                                 "Collection was modified? while trying to save game db, will loop and try to save again! [Failures: " +
                                 failures +
                                 "]"
@@ -1469,14 +1465,10 @@ namespace Intersect.Server.Database
                                 "Error Saving Game Database! Server will shutdown in order to prevent potential rollback scenarios!"
                             );
 
-                            Task.Factory.StartNew(
-                                () => Bootstrapper.OnUnhandledException(
-                                    Thread.CurrentThread.Name, new UnhandledExceptionEventArgs(ex, true)
-                                )
-                            );
+                            ServerContext.DispatchUnhandledException(exception);
 
                             gameDbLogger?.Error(
-                                ex,
+                                exception,
                                 "Error Saving Game Database! Server will shutdown in order to prevent potential rollback scenarios!"
                             );
 
@@ -1566,9 +1558,9 @@ namespace Intersect.Server.Database
                                 : $"{currentTrace.Id:00000}: Save complete but there are no available traces."
                         );
                     }
-                    catch (Exception ex)
+                    catch (Exception exception)
                     {
-                        if (ex is InvalidOperationException)
+                        if (exception is InvalidOperationException)
                         {
                             //Collection was modified?
                             //Loop and try to save again!
@@ -1579,14 +1571,10 @@ namespace Intersect.Server.Database
                                     "Error Saving Player Database! Server will shutdown in order to prevent potential rollback scenarios!"
                                 );
 
-                                Task.Factory.StartNew(
-                                    () => Bootstrapper.OnUnhandledException(
-                                        Thread.CurrentThread.Name, new UnhandledExceptionEventArgs(ex, true)
-                                    )
-                                );
+                                ServerContext.DispatchUnhandledException(exception);
 
                                 playerDbLogger?.Error(
-                                    ex,
+                                    exception,
                                     "Error Saving Player Database! Server will shutdown in order to prevent potential rollback scenarios! [Failures: " +
                                     failures +
                                     "]"
@@ -1598,7 +1586,7 @@ namespace Intersect.Server.Database
                             //This should be a warning but I want to actually see it working in a real environment without people needing to change their configs for a few builds
                             //TODO change to .Warn
                             playerDbLogger?.Error(
-                                ex,
+                                exception,
                                 "Collection was modified? while trying to save player db, will loop and try to save again! [Failures: " +
                                 failures +
                                 "]"
@@ -1610,14 +1598,10 @@ namespace Intersect.Server.Database
                                 "Error Saving Player Database! Server will shutdown in order to prevent potential rollback scenarios!"
                             );
 
-                            Task.Factory.StartNew(
-                                () => Bootstrapper.OnUnhandledException(
-                                    Thread.CurrentThread.Name, new UnhandledExceptionEventArgs(ex, true)
-                                )
-                            );
+                            ServerContext.DispatchUnhandledException(exception);
 
                             playerDbLogger?.Error(
-                                ex,
+                                exception,
                                 "Error Saving Player Database! Server will shutdown in order to prevent potential rollback scenarios!"
                             );
 
@@ -1881,7 +1865,7 @@ namespace Intersect.Server.Database
             }
 
             Console.WriteLine(Strings.Migration.migrationcomplete);
-            Bootstrapper.Context.ServerConsole.Wait(true);
+            Bootstrapper.Context.ConsoleService.Wait(true);
             Environment.Exit(0);
         }
